--- strelets/remove_virtual_chain.go
@@ -1,13 +1,0 @@
-package strelets
-
-import (
-	"context"
-)
-
-type RemoveVirtualChainInput struct {
-	VirtualChain *VirtualChain
-}
-
-func (s *strelets) RemoveVirtualChain(ctx context.Context, input *RemoveVirtualChainInput) error {
-	panic("removed")
-}
