--- strelets/shared_network.go
@@ -1,13 +1,0 @@
-package strelets
-
-import (
-	"context"
-)
-
-type ProvisionSharedNetworkInput struct {
-	Name string
-}
-
-func (s *strelets) ProvisionSharedNetwork(ctx context.Context, input *ProvisionSharedNetworkInput) error {
-	panic("removed")
-}
